RoleAllocation: isolate specific BruteForce algorithm and prepare to switch to different algorithm (like ID-based, position-based)
